requirements.txt
numpy==1.26.4
sympy~=1.12
matplotlib==3.8.3
numba==0.59.0
scipy==1.12.0
Pillow==10.2.0
ffmpeg-python
mypy
fastapi~=0.110.0
uvicorn[standard]~=0.27.1
pydantic~=2.6.3
pytest~=8.1.1
trio